backend/pyproject.toml
[project]
name = "youtube-empire-backend"
version = "0.1.0"
description = "Agentic AI Framework — Multi-Agent Communication Platform with Tool Execution"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "python-frontmatter>=1.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "python-socketio>=5.11.0",
    "pyyaml>=6.0.0",
]
